Adds missing author, license, and copyright_holder to dist.ini.
If author, license, or copyright_holder are missing from dist.ini, dzil
fall backs to the information in the users ~/.dzil/config.ini. If the
user is not the author of the module, it will cause README.md to be
populated with the wrong author information. If the user does not have
a ~/.dzil/config.ini file, "dzil build" will fail due to missing author
and license.